feat: Windows support, mingw-w64 toolchain, Win32-native platform layer
Third supported host: Windows 11 x86-64. The compiler drives LLVM's clang
with --target=x86_64-w64-mingw32 against an MSYS2 UCRT sysroot, so compiled
programs need no extra runtime. The platform semantics are written against
Win32 with Node's observable behaviour as the reference, not through a
POSIX emulation layer.

Toolchain (ADR-00718)
- hostclang.go: single host-specific clang argv (target, sysroot,
  KLAIN_SYSROOT override, .exe suffix) shared by the driver, the E2E
  helpers, and the conformance runner.
- LLP64 fixes in the embedded C (long → int64_t), curl's 32-bit long
  outputs, zlib's z_stream layout, and a 16-byte-aligned _setjmp buffer.
- .gitattributes pins LF; .gitignore covers the Windows binaries.

Platform layer (ADR-00719) — four shim sources compiled once per content
hash into %LOCALAPPDATA%\klainmain\win32shim and linked into every program:
- win32shim.c: getline, dprintf, setenv, sysconf, getrandom (BCrypt),
  clock_gettime/nanosleep on the IR's 64-bit timespec, isatty via the
  console API, rss, os.cpus() from the registry and processor times, and
  no Windows Error Reporting dialog on a crash.
- win32io.c: fd table with a socket range that never touches the CRT,
  Winsock select reactor (WSAPoll never reports a failed connect),
  reset→EOF emulation, shutdown-before-close, Win32 fibers behind
  getcontext/makecontext/swapcontext, and a curl_multi_fdset bridge.
- win32fs.c: glibc stat layout with birthtime, Node mode synthesis,
  UTF-16 paths, own dirent, 64-bit seek, Linux errno numbers with libuv
  wording, symlink/readlink/realpath.
- win32proc.c: CreateProcessW spawn with libuv quoting, pid table for
  waitpid/kill, socketpair over loopback, inherited-socket handshake for
  child_process.fork/cluster, console control handlers for signals.
- Emitter hooks: os module (USERPROFILE, TEMP, \r\n), fs errno/stat/dirent
  layouts, addrinfo and ucontext layouts, fork sites rewritten to spawn
  helpers, LoadLibrary for lazy import() islands, console-mode tty/tui,
  ETIMEDOUT on winpthreads, no GC_set_handle_fork, goroutine scheduler
  thread-locals read through out-of-line accessors.
- TLS: OpenSSL is handed the Winsock SOCKET (not the fd) and the
  non-blocking read sets EAGAIN itself — tls, https, secure http2, wss.

Cross-platform bugs found by the port, each with an E2E test
- process.argv's headerized array was not NULL-terminated (execv read
  past the end on every host).
- The first .then on a fetch typed its callback parameter as number.
- Null reads LLVM lowers to a crash on Linux and a self-loop on Windows:
  new Set().has(null), parseInt/parseFloat of null or undefined,
  new RegExp(p, undefined).
- Two tests handed the parser's output straight to the emitter with a
  top-level function named like a C symbol (read, main); they now go
  through the resolver, which mangles it. The main one had been a masked
  skip everywhere.

Tests and tooling
- Windows variants and documented skips: signal delivery, POSIX-only
  tools, sanitizers, symlinks without Developer Mode, getuid, cluster
  worker distribution (shared accept instead of round-robin).
- A pure-Go loopback probe skips network tests when a traffic filter
  (AdGuard on the dev box) is intercepting 127.0.0.1.
- Conformance runner builds with the host clang argv and writes
  slash-separated report paths.

Verified on Windows 11 x86-64 (2026-09-06): go test ./tests 2636 passed,
0 failed, 47 skipped; 219 of 222 examples (the rest platform-expected);
Test262 strict 6260 / js 6674, six above the Linux baseline per lane.
Committed conformance reports remain the Linux ones.

Known gaps, documented in TDD-00177 and the README: path stays
POSIX-shaped (no path.win32 default), cluster round-robin, --static,
-package, Webview, -crypto=commoncrypto, and no ASan/UBSan runtime for
the mingw target.

Docs: README Windows install section and PATH notes, docs/tdd/TDD-00177,
docs/adr/ADR-00718 and ADR-00719, status data and regenerated pages.

+++ codegen/llvm/buffersrc/buffer_codecs.c
@@ -8,6 +8,7 @@
  * Allocation goes through plain malloc, so under -mm=gc it is transparently
  * routed to the collector by the shim's global malloc override.
  */
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -16,19 +17,19 @@
  * base+8, so KML string ops read the true length via ptr-8. The codec encoders
  * (hex/base64/latin1 -> string) return one of these; the decoders (-> byte
  * buffer) keep plain malloc. */
-static char *kml_str_hdr_alloc(long len) {
+static char *kml_str_hdr_alloc(int64_t len) {
 	char *base = (char *)malloc((size_t)(8 + len + 1));
 	if (!base) return base;
-	*(long *)base = len;
+	*(int64_t *)base = len;
 	return base + 8;
 }
 
 /* ---- hex ---- */
 
-char *__kml_buf_hex_enc(const unsigned char *src, long long n) {
+char *__kml_buf_hex_enc(const unsigned char *src, int64_t n) {
 	static const char digits[] = "0123456789abcdef";
 	char *out = kml_str_hdr_alloc(n * 2);
-	for (long long i = 0; i < n; i++) {
+	for (int64_t i = 0; i < n; i++) {
 		out[i * 2] = digits[src[i] >> 4];
 		out[i * 2 + 1] = digits[src[i] & 0xF];
 	}
@@ -45,10 +46,10 @@ static int hexval(char c) {
 
 /* Node semantics: decoding stops at the first non-hex pair (and a trailing
  * lone digit is dropped). Returns the byte count; *out receives the buffer. */
-long long __kml_buf_hex_dec(const char *s, unsigned char **out) {
+int64_t __kml_buf_hex_dec(const char *s, unsigned char **out) {
 	size_t sl = strlen(s);
 	unsigned char *buf = (unsigned char *)malloc(sl / 2 + 1);
-	long long n = 0;
+	int64_t n = 0;
 	for (size_t i = 0; i + 1 < sl; i += 2) {
 		int hi = hexval(s[i]), lo = hexval(s[i + 1]);
 		if (hi < 0 || lo < 0) break;
@@ -60,14 +61,14 @@ long long __kml_buf_hex_dec(const char *s, unsigned char **out) {
 
 /* ---- base64 / base64url ---- */
 
-char *__kml_buf_b64_enc(const unsigned char *src, long long n, int urlsafe) {
+char *__kml_buf_b64_enc(const unsigned char *src, int64_t n, int urlsafe) {
 	static const char std_al[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
 	static const char url_al[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_";
 	const char *al = urlsafe ? url_al : std_al;
-	long long groups = (n + 2) / 3;
+	int64_t groups = (n + 2) / 3;
 	char *out = kml_str_hdr_alloc(groups * 4);
-	long long o = 0;
-	for (long long i = 0; i < n; i += 3) {
+	int64_t o = 0;
+	for (int64_t i = 0; i < n; i += 3) {
 		unsigned v = (unsigned)src[i] << 16;
 		if (i + 1 < n) v |= (unsigned)src[i + 1] << 8;
 		if (i + 2 < n) v |= (unsigned)src[i + 2];
@@ -93,10 +94,10 @@ static int b64val(char c) {
 
 /* Lenient like Node: non-alphabet bytes (whitespace, '=') are skipped;
  * both the standard and url-safe alphabets are accepted. */
-long long __kml_buf_b64_dec(const char *s, unsigned char **out) {
+int64_t __kml_buf_b64_dec(const char *s, unsigned char **out) {
 	size_t sl = strlen(s);
 	unsigned char *buf = (unsigned char *)malloc(sl / 4 * 3 + 3);
-	long long n = 0;
+	int64_t n = 0;
 	unsigned acc = 0;
 	int bits = 0;
 	for (size_t i = 0; i < sl; i++) {
@@ -117,10 +118,10 @@ long long __kml_buf_b64_dec(const char *s, unsigned char **out) {
 
 /* bytes -> UTF-8 string: 0x00–0x7F pass through, 0x80–0xFF become the
  * 2-byte UTF-8 encoding of U+0080–U+00FF. */
-char *__kml_buf_latin1_str(const unsigned char *src, long long n) {
+char *__kml_buf_latin1_str(const unsigned char *src, int64_t n) {
 	char *out = kml_str_hdr_alloc(n * 2); /* max; actual length set below */
-	long long o = 0;
-	for (long long i = 0; i < n; i++) {
+	int64_t o = 0;
+	for (int64_t i = 0; i < n; i++) {
 		unsigned char b = src[i];
 		if (b < 0x80) {
 			out[o++] = (char)b;
@@ -130,16 +131,16 @@ char *__kml_buf_latin1_str(const unsigned char *src, long long n) {
 		}
 	}
 	out[o] = 0;
-	*(long *)(out - 8) = o; /* TDD-00120: actual UTF-8 length */
+	*(int64_t *)(out - 8) = o; /* TDD-00120: actual UTF-8 length */
 	return out;
 }
 
 /* UTF-8 string -> bytes: each codepoint keeps its low 8 bits (Node's
  * latin1-write masking). Invalid sequences fall back byte-wise. */
-long long __kml_buf_latin1_bytes(const char *s, unsigned char **out) {
+int64_t __kml_buf_latin1_bytes(const char *s, unsigned char **out) {
 	size_t sl = strlen(s);
 	unsigned char *buf = (unsigned char *)malloc(sl + 1);
-	long long n = 0;
+	int64_t n = 0;
 	for (size_t i = 0; i < sl;) {
 		unsigned char c = (unsigned char)s[i];
 		unsigned cp = c;
@@ -165,24 +166,24 @@ long long __kml_buf_latin1_bytes(const char *s, unsigned char **out) {
  * base64 the DER in 64-char lines between -----BEGIN <label>-----/
  * -----END <label>----- with a trailing newline — Node's own PEM shape.
  * Returns a length-prefixed kml string. */
-char *__kml_pem_from_der(const unsigned char *der, long long n, const char *label) {
+char *__kml_pem_from_der(const unsigned char *der, int64_t n, const char *label) {
 	char *b64 = __kml_buf_b64_enc(der, n, 0);
-	long long bl = *(long *)(b64 - 8);
-	long long lines = (bl + 63) / 64;
-	long long ll = (long long)strlen(label);
+	int64_t bl = *(int64_t *)(b64 - 8);
+	int64_t lines = (bl + 63) / 64;
+	int64_t ll = (int64_t)strlen(label);
 	/* "-----BEGIN -----\n" = 17 + label; "-----END -----\n" = 15 + label */
-	long long total = 17 + ll + bl + lines + 15 + ll;
+	int64_t total = 17 + ll + bl + lines + 15 + ll;
 	char *out = kml_str_hdr_alloc(total);
-	long long o = 0;
+	int64_t o = 0;
 	o += sprintf(out + o, "-----BEGIN %s-----\n", label);
-	for (long long i = 0; i < bl; i += 64) {
-		long long chunk = bl - i < 64 ? bl - i : 64;
+	for (int64_t i = 0; i < bl; i += 64) {
+		int64_t chunk = bl - i < 64 ? bl - i : 64;
 		memcpy(out + o, b64 + i, (size_t)chunk);
 		o += chunk;
 		out[o++] = '\n';
 	}
 	o += sprintf(out + o, "-----END %s-----\n", label);
-	*(long *)(out - 8) = o;
+	*(int64_t *)(out - 8) = o;
 	out[o] = 0;
 	return out;
 }
